Use mktemp -t tmp.XXXXXXXXXX to support running on OS X
Running mktemp without any arguments on OS X causes an error message to
be returned. Adding the argument -t tmp.XXXXXXXXXX will support running
on OS X and Linux.

This is only tested on OS X and Ubuntu, and according to the man pages
on Ubuntu, the -t flag is deprecated, but doesn't seem to have any ill effect.